+++ src/main/java/module-info.java
@@ -9,4 +9,6 @@ module uni.usic.taskmanagerv1javafx {
     exports uni.usic.taskmanager;
     exports uni.usic.taskmanager.views.tasks;
     opens uni.usic.taskmanager.views.tasks to javafx.fxml;
+    exports uni.usic.taskmanager.views.users;
+    opens uni.usic.taskmanager.views.users to javafx.fxml;
 }
